Move controls into own file
